Hide waitForEntitlement from public API
