+++ src/main.ts
@@ -24,6 +24,7 @@ import {
   CustomerInfo as InnerCustomerInfo,
   toCustomerInfo,
 } from "./entities/customer-info";
+import { waitForEntitlement } from "./helpers/entitlement-checking-helper";
 
 export type Offerings = InnerOfferings;
 export type Offering = InnerOffering;
@@ -180,35 +181,6 @@ export class Purchases {
     return entitlements.includes(entitlementIdentifier);
   }
 
-  public waitForEntitlement(
-    appUserId: string,
-    entitlementIdentifier: string,
-    maxAttempts: number = 10,
-  ): Promise<boolean> {
-    const waitMSBetweenAttempts = 1000;
-    return new Promise<boolean>((resolve, reject) => {
-      const checkForEntitlement = (checkCount = 1) =>
-        this.isEntitledTo(appUserId, entitlementIdentifier)
-          .then((hasEntitlement) => {
-            if (checkCount > maxAttempts) {
-              return resolve(false);
-            }
-
-            if (hasEntitlement) {
-              return resolve(true);
-            } else {
-              setTimeout(
-                () => checkForEntitlement(checkCount + 1),
-                waitMSBetweenAttempts,
-              );
-            }
-          })
-          .catch(reject);
-
-      checkForEntitlement();
-    });
-  }
-
   public async subscribe(
     appUserId: string,
     productId: string,
@@ -304,7 +276,8 @@ export class Purchases {
           environment,
           customerEmail,
           onFinished: async () => {
-            const hasEntitlement = await this.waitForEntitlement(
+            const hasEntitlement = await waitForEntitlement(
+              this,
               appUserId,
               entitlementId,
             );
